Add return type of MuxApplication
This change allows to run mux application that returns any type.  In
consenus there is an application that returns a Void, this way we don't
need to change it to run on the mux layer.

In the future we could also change mux layer so that one can access the
result term.