Update to cardano-sl version with OBFT EBB epoch 0 fix
When we start a chain from a genesis in the OBFT era, rather than the
Original era, the epoch zero EBB must still exist, even though all other
epoch EBBs do not exist.

That EBB ought to have an empty slot leader schedule, otherwise we
cannot have new non-legcay nodes create the same EBB, since new nodes do
not implement the logic for Original era leader schedules. This does not
affects any existing chain, since existing chains all start in the
Original era and go through a hard fork to enter the OBFT era. This
change only affects new test/dev chains that start in the OBFT era. This
is needed to test mixed clusters of old/new nodes in a convenient way,
that all start from genesis in the OBFT era.

This patch updates to a version of cardano-sl that has the necessary
changes to do this. The proxy also has to put the initial EBB into its
local chain database to be able to talk to old and new nodes.